Moving async_strem_writer to postgres package to avoid package name with underscore (qlik_connect)
