+++ postgres/async_stream_writer.go
@@ -1,21 +1,25 @@
-package qlik_connect
+package postgres
+
+import (
+	qlik "github.com/qlik-ea/postgres-grpc-connector/qlik"
+)
 
 // AsyncStreamWriter defines the writer interface.
 type AsyncStreamWriter struct {
-	grpcStream Connector_GetDataServer
-	channel    chan *DataChunk
+	grpcStream qlik.Connector_GetDataServer
+	channel    chan *qlik.DataChunk
 	done       chan bool
 }
 
 // NewAsyncStreamWriter constructs a new async stream writer.
-func NewAsyncStreamWriter(grpcStream Connector_GetDataServer, done chan bool) *AsyncStreamWriter {
-	var this = &AsyncStreamWriter{grpcStream, make(chan *DataChunk, 10), done}
+func NewAsyncStreamWriter(grpcStream qlik.Connector_GetDataServer, done chan bool) *AsyncStreamWriter {
+	var this = &AsyncStreamWriter{grpcStream, make(chan *qlik.DataChunk, 10), done}
 	go this.run()
 	return this
 }
 
 // Write will send a datachunk on the underlying stream.
-func (a *AsyncStreamWriter) Write(rowBundle *DataChunk) {
+func (a *AsyncStreamWriter) Write(rowBundle *qlik.DataChunk) {
 	a.channel <- rowBundle
 }
